+++ src/collective/cover/tiles/contentbody.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from AccessControl import Unauthorized
+from collective.cover import _
 from collective.cover.tiles.base import IPersistentCoverTile
 from collective.cover.tiles.base import PersistentCoverTile
 from plone.app.uuid.utils import uuidToObject
@@ -19,10 +20,11 @@ class ContentBodyTile(PersistentCoverTile):
 
     implements(IPersistentCoverTile)
 
-    index = ViewPageTemplateFile("templates/contentbody.pt")
+    index = ViewPageTemplateFile('templates/contentbody.pt')
 
     is_editable = False
     is_configurable = False
+    short_name = _(u'msg_short_name_contentbody', default=u'Content Body')
 
     def body(self):
         body = ''
@@ -50,6 +52,7 @@ class ContentBodyTile(PersistentCoverTile):
         data_mgr.set(data)
 
     def accepted_ct(self):
-        """ For now we are supporting Document and News Item
+        """Return 'Document' and 'News Item' as the only content types
+        accepted in the tile.
         """
         return ['Document', 'News Item']