+++ src/collective/cover/tiles/pfg.py
@@ -33,10 +33,11 @@ class PFGTile(PersistentCoverTile):
 
     implements(IPersistentCoverTile)
 
-    index = ViewPageTemplateFile("templates/pfg.pt")
+    index = ViewPageTemplateFile('templates/pfg.pt')
 
     is_editable = True
     is_configurable = True
+    short_name = _(u'msg_short_name_pfg', default=u'FormGen')
 
     def body(self):
         body = ''
@@ -62,6 +63,6 @@ class PFGTile(PersistentCoverTile):
         data_mgr.set(data)
 
     def accepted_ct(self):
-        """ For now we are supporting Document and News Item
+        """Return 'FormFolder' as the only content type accepted in the tile.
         """
         return ['FormFolder']